+++ src/test/java/com/example/moneyAllocation/repository/AccountRepositoryImplTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertThrowsExactly;
 import com.example.moneyAllocation.domain.Account;
 import com.example.moneyAllocation.domain.AccountSelector;
 import com.example.moneyAllocation.exception.ResourceNotFoundException;
+import com.example.moneyAllocation.repository.impl.AccountRepositoryImpl;
 import com.example.moneyAllocation.repository.mybatis.AccountMapper;
 import java.util.ArrayList;
 import java.util.List;
