+++ src/test/java/com/example/moneyAllocation/repository/TemporaryTransferRepositoryImplTest.java
@@ -5,6 +5,7 @@ import com.example.moneyAllocation.domain.TemporaryTransfer;
 import com.example.moneyAllocation.domain.TemporaryTransferSelector;
 import com.example.moneyAllocation.domain.TransferSelector;
 import com.example.moneyAllocation.exception.ResourceNotFoundException;
+import com.example.moneyAllocation.repository.impl.TemporaryTransferRepositoryImpl;
 import com.example.moneyAllocation.repository.mybatis.TemporaryTransferMapper;
 import java.util.ArrayList;
 import java.util.List;
