+++ src/test/java/com/example/moneyAllocation/service/TemporaryTransferServiceImplTest.java
@@ -10,6 +10,7 @@ import com.example.moneyAllocation.exception.ResourceNotFoundException;
 import com.example.moneyAllocation.exception.ResourceValidationException;
 import com.example.moneyAllocation.repository.AccountRepository;
 import com.example.moneyAllocation.repository.TemporaryTransferRepository;
+import com.example.moneyAllocation.service.impl.TemporaryTransferServiceImpl;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.AfterEach;
